refactor: split probe-scenario.ts (873 lines) into 5 focused modules
Split queue handler into:
- probe-scenario/handler.ts (391) — main job processor
- probe-scenario/retry.ts (237) — retry logic + error handling
- probe-scenario/worker-input.ts (211) — scenario fetching + input building
- probe-scenario/health-check.ts (102) — Python worker health check
- probe-scenario/index.ts (10) — barrel

Codex: 3/5 PASS (2 false alarms on pre-existing handler index size)
