+++ cloud/apps/api/src/queue/handlers/handler-config.ts
@@ -1,0 +1,117 @@
+/**
+ * Handler configuration — queue names, types, and registration table.
+ * Extracted from index.ts to keep it under the 400-line file size limit.
+ */
+
+import type { PgBoss } from 'pg-boss';
+import type {
+  JobType,
+  ProbeScenarioJobData,
+  SummarizeTranscriptJobData,
+  AnalyzeBasicJobData,
+  ExpandScenariosJobData,
+  ComputeTokenStatsJobData,
+  ProbeDeadLetterJobData,
+  AggregateAnalysisJobData,
+} from '../types.js';
+import { createProbeScenarioHandler } from './probe-scenario/index.js';
+import { createSummarizeTranscriptHandler } from './summarize-transcript.js';
+import { createAnalyzeBasicHandler } from './analyze-basic.js';
+import { createExpandScenariosHandler } from './expand-scenarios.js';
+import { createComputeTokenStatsHandler } from './compute-token-stats.js';
+import { createProbeDeadLetterHandler } from './probe-dead-letter.js';
+import { createAggregateAnalysisHandler } from './aggregate-analysis.js';
+
+// Re-export job data types for handlers
+export type {
+  ProbeScenarioJobData,
+  SummarizeTranscriptJobData,
+  AnalyzeBasicJobData,
+  ExpandScenariosJobData,
+  ComputeTokenStatsJobData,
+  ProbeDeadLetterJobData,
+  AggregateAnalysisJobData,
+};
+
+// Dead letter queue name for probe jobs
+export const PROBE_DEAD_LETTER_QUEUE = 'probe_dead_letter';
+
+// Handler registration info
+export type HandlerRegistration = {
+  name: JobType;
+  register: (boss: PgBoss, batchSize: number) => Promise<void>;
+};
+
+export const handlerRegistrations: HandlerRegistration[] = [
+  {
+    name: 'probe_scenario',
+    register: async (boss, batchSize) => {
+      // Register for the default probe_scenario queue (fallback)
+      await boss.work<ProbeScenarioJobData>(
+        'probe_scenario',
+        { batchSize },
+        createProbeScenarioHandler()
+      );
+    },
+  },
+  {
+    name: 'summarize_transcript',
+    register: async (boss, batchSize) => {
+      await boss.work<SummarizeTranscriptJobData>(
+        'summarize_transcript',
+        { batchSize },
+        createSummarizeTranscriptHandler()
+      );
+    },
+  },
+  {
+    name: 'analyze_basic',
+    register: async (boss, batchSize) => {
+      await boss.work<AnalyzeBasicJobData>(
+        'analyze_basic',
+        { batchSize },
+        createAnalyzeBasicHandler()
+      );
+    },
+  },
+  {
+    name: 'expand_scenarios',
+    register: async (boss, batchSize) => {
+      await boss.work<ExpandScenariosJobData>(
+        'expand_scenarios',
+        { batchSize },
+        createExpandScenariosHandler()
+      );
+    },
+  },
+  {
+    name: 'compute_token_stats',
+    register: async (boss, batchSize) => {
+      await boss.work<ComputeTokenStatsJobData>(
+        'compute_token_stats',
+        { batchSize },
+        createComputeTokenStatsHandler()
+      );
+    },
+  },
+  {
+    name: 'probe_dead_letter',
+    register: async (boss, batchSize) => {
+      await boss.work<ProbeDeadLetterJobData>(
+        PROBE_DEAD_LETTER_QUEUE,
+        { batchSize },
+        createProbeDeadLetterHandler()
+      );
+    },
+  },
+  {
+    name: 'aggregate_analysis',
+    register: async (boss, batchSize) => {
+      await boss.work<AggregateAnalysisJobData>(
+        'aggregate_analysis',
+        { batchSize }, // Usually batchSize=1 effectively for aggregation if we want strict serial per worker
+        createAggregateAnalysisHandler()
+      );
+    },
+  },
+];
