+++ cloud/apps/api/src/queue/handlers/probe-scenario/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Probe Scenario Handler — barrel export.
+ *
+ * Public API surface mirrors the old probe-scenario.ts module so that
+ * callers need only update the import path.
+ */
+
+export { createProbeScenarioHandler } from './handler.js';
+export { resetHealthCheck, ensureHealthCheck } from './health-check.js';
+export { isRetryableError } from './retry.js';
